+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 - add switch to toggle simultaneous account syncing off
 - Implement settings for socks5 proxy
 
+## Changed
+- hide unfinished themes with the prefix `dev_` from the theme selection, unless `--devmode` is active.
+- hide rocket theme
+
 ### Fixed
 - fix flashing up account list on startup
 - fix update/load core translation strings